DAOS-2828 control: Refactor structured fault type
It was a mistake to name the package "faults"; fixing that
before it gets wider use. Also moving codes into a subpackage
as it reads better (e.g. code.StorageAlreadyFormatted) and
should be more maintainable as we add more codes.
